changed needle length type to size_t in kore_mem_find(). updated function prototype in kore.h. Use tab instead of space indentation

+++ src/utils.c
@@ -482,11 +482,11 @@ kore_base64_decode(char *in, u_int8_t **out, u_int32_t *olen)
 }
 
 void *
-kore_mem_find(void *src, size_t slen, void *needle, u_int32_t len)
+kore_mem_find(void *src, size_t slen, void *needle, size_t len)
 {
-  size_t  pos;
+	size_t  pos;
 
-  for(pos = 0; pos < slen; pos++) {
+	for(pos = 0; pos < slen; pos++) {
 		if ( *((u_int8_t *)src + pos) != *(u_int8_t *)needle)
 			continue;
